+++ lib/src/motion_ticker.dart
@@ -1,5 +1,13 @@
 part of just_motion;
 
+/// [TickerMan] sits at the core of **just_motion**, and manages the single
+/// [Ticker] that runs all `MotionValue` simulations.
+/// Detects when a [MotionValue] subscribes (meaning the `value!=target`)
+/// and starts the ticker if not running; when all
+/// motions are completed, the ticker stops.
+///
+/// [TickerMan] shouldn't be commonly used externally by the developer.
+/// as it manages the state internally with [MotionValue].
 class TickerMan {
   static TickerMan _instance = TickerMan();
 
@@ -10,14 +18,17 @@ class TickerMan {
 
   void isActive() => _ticker.isActive;
 
+  /// Stops the the `Ticker` provider
   void stop() {
     if (_ticker.isActive) _ticker.stop();
   }
 
+  /// Starts the the `Ticker` provider
   void start() {
     if (!_ticker.isActive) _ticker.start();
   }
 
+  /// callback that runs the simulation on any active [MotionValue].
   void _onTick(Duration t) {
     final remover = <MotionValue>{};
     final _safeMotions = Set.of(_motions);
@@ -35,13 +46,15 @@ class TickerMan {
     }
   }
 
-  void remove<T>(MotionValue<T> euler) {
-    _motions.remove(euler);
+  /// Removes a [MotionValue] object from the ticking.
+  void remove<T>(MotionValue<T> motion) {
+    _motions.remove(motion);
   }
 
-  void activate<T>(MotionValue<T> euler) {
-    if (!euler.completed) {
-      _motions.add(euler);
+  /// Activates a [MotionValue] object and adds it to the ticking queue.
+  void activate<T>(MotionValue<T> motion) {
+    if (!motion.completed) {
+      _motions.add(motion);
       start();
     }
   }